apps/backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.110,<1.0
uvicorn[standard]>=0.24,<1.0

# Database
sqlalchemy>=2.0,<3.0
alembic>=1.13,<2.0
psycopg2-binary>=2.9,<3.0  # PostgreSQL adapter

# Authentication and Security
pydantic[email]>=2.0,<3.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
python-multipart>=0.0.6

# Supabase integration
supabase>=2.0,<3.0

# Redis for caching
redis>=5.0,<6.0
celery>=5.3,<6.0

# File handling and storage
boto3>=1.34,<2.0  # AWS S3
python-magic>=0.4,<1.0  # File type detection

# HTTP client for external services
httpx>=0.27,<1.0

# WebSocket support
websockets>=12.0,<13.0

# Development and testing
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
pytest-cov>=4.0,<5.0
factory-boy>=3.3,<4.0

# Utilities
python-dotenv>=1.0,<2.0
pydantic-settings>=2.0,<3.0

# Security
bleach>=6.0,<7.0
python-magic>=0.4,<1.0

# Performance monitoring
psutil>=5.9,<6.0
